+++ commands/fun/lewd.js
@@ -1,22 +1,26 @@
 module.exports = {
-  name: "neko",
+  name: "lewd",
   category: "fun",
-  description: "See a randomly generated picture of a Neko, also known as a Catgirl.",
+  description: "A randomly generated photo of *you know what*",
   usage: "",
   example: "",
-  options: { permission: "EVERYONE", enabled: true, cooldown: 3, guildOnly: false },
+  options: { permission: "EVERYONE", enabled: true, cooldown: 3, guildOnly: true },
   run: async (bot, message, args) => {
     
     const Discord = require("discord.js");
     const fetch = require("node-fetch");
+
+    if(message.channel.nsfw === false) {
+      return message.reply("Run this command in an **NSFW** channel.")
+    }
 
     let url;
     await fetch(`http://api.nekos.fun:8080/api/lewd`)
         .then(res => res.json())
         .then(json => url = json.image)
 
-    const nekoEmbed = {
-      "title": "I present your Neko:",
+    const lewdEmbed = {
+      "title": "Here you go, you filthy animal:",
       "image": {
         "url": url
       },
@@ -26,7 +30,7 @@ module.exports = {
       }
     }
 
-    message.reply({ embeds: [ nekoEmbed ]});
+    message.reply({ embeds: [ lewdEmbed ]});
 
   },
 }